cleancode database migration

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,32 +13,34 @@ class CreateUsersTable extends Migration
    */
   public function up()
   {
-    Schema::create("users", function (Blueprint $table) {
-      $table->bigIncrements("id");
-      $table->string("name");
-      $table->string("email")->unique();
-      $table->timestamp("email_verified_at")->nullable();
-      $table->string("password");
-      $table->rememberToken();
-      $table->timestamps();
-      $table
-        ->string("role", 100)
-        ->comment("superadmin, admin, pendaftaran, poli, channel")
-        ->nullable();
-      $table->char("poli", 10)->nullable();
-      $table
-        ->char("loket", 10)
-        ->comment("nomor loket")
-        ->nullable();
-      $table
-        ->char("loket_lansia", 1)
-        ->default("0")
-        ->nullable();
-      $table
-        ->char("prioritas", 1)
-        ->default("0")
-        ->nullable();
-    });
+    if (!Schema::hasTable('users')) {
+      Schema::create("users", function (Blueprint $table) {
+        $table->bigIncrements("id");
+        $table->string("name");
+        $table->string("email")->unique();
+        $table->timestamp("email_verified_at")->nullable();
+        $table->string("password");
+        $table->rememberToken();
+        $table->timestamps();
+        $table
+          ->string("role", 100)
+          ->comment("superadmin, admin, pendaftaran, poli, channel")
+          ->nullable();
+        $table->char("poli", 10)->nullable();
+        $table
+          ->char("loket", 10)
+          ->comment("nomor loket")
+          ->nullable();
+        $table
+          ->char("loket_lansia", 1)
+          ->default("0")
+          ->nullable();
+        $table
+          ->char("prioritas", 1)
+          ->default("0")
+          ->nullable();
+      });
+    }
   }
 
   /**
@@ -48,6 +50,8 @@ class CreateUsersTable extends Migration
    */
   public function down()
   {
-    Schema::dropIfExists("users");
+    if (Schema::hasTable('users')) {
+      Schema::dropIfExists("users");
+    }
   }
 }
