+++ database/migrations/2019_12_24_102620_create_table_printer.php
@@ -13,20 +13,22 @@ class CreateTablePrinter extends Migration
    */
   public function up()
   {
-    Schema::create("ant_printer", function (Blueprint $table) {
-      $table->bigIncrements("printer_id");
-      $table->string("printer_nama", 100);
-      $table
-        ->string("printer_url_service", 225)
-        ->comment("URL Service untuk print");
-      $table
-        ->string("printer_alias", 100)
-        ->default("epson")
-        ->nullable();
-      $table->string("printer_modifier", 100)->nullable();
-      $table->dateTime("printer_modified")->nullable();
-      $table->string("printer_code", 100);
-    });
+    if (!Schema::hasTable('ant_printer')) {
+      Schema::create("ant_printer", function (Blueprint $table) {
+        $table->bigIncrements("printer_id");
+        $table->string("printer_nama", 100);
+        $table
+          ->string("printer_url_service", 225)
+          ->comment("URL Service untuk print");
+        $table
+          ->string("printer_alias", 100)
+          ->default("epson")
+          ->nullable();
+        $table->string("printer_modifier", 100)->nullable();
+        $table->dateTime("printer_modified")->nullable();
+        $table->string("printer_code", 100);
+      });
+    }
   }
 
   /**
@@ -36,6 +38,8 @@ class CreateTablePrinter extends Migration
    */
   public function down()
   {
-    Schema::dropIfExists("ant_printer");
+    if (Schema::hasTable('ant_printer')) {
+      Schema::dropIfExists("ant_printer");
+    }
   }
 }
