+++ database/migrations/2021_06_23_140459_create_ant_service_table.php
@@ -13,18 +13,20 @@ class CreateAntServiceTable extends Migration
    */
   public function up()
   {
-    Schema::create("ant_service", function (Blueprint $table) {
-      $table->bigIncrements("service_id");
-      $table->string("service_nama", 100);
-      $table->string("service_url", 225)->comment("URL Service");
-      $table->string("service_modifier", 100)->nullable();
-      $table->dateTime("service_modified")->nullable();
-      $table->string("service_code", 100);
-    });
+    if (!Schema::hasTable('ant_service')) {
+      Schema::create("ant_service", function (Blueprint $table) {
+        $table->bigIncrements("service_id");
+        $table->string("service_nama", 100);
+        $table->string("service_url", 225)->comment("URL Service");
+        $table->string("service_modifier", 100)->nullable();
+        $table->dateTime("service_modified")->nullable();
+        $table->string("service_code", 100);
+      });
 
-    DB::statement(
-      "ALTER Table ant_service add service_enabled INT(1) NULL DEFAULT 0 after service_code"
-    );
+      DB::statement(
+        "ALTER Table ant_service add service_enabled INT(1) NULL DEFAULT 0 after service_code"
+      );
+    }
   }
 
   /**
@@ -34,6 +36,8 @@ class CreateAntServiceTable extends Migration
    */
   public function down()
   {
-    Schema::dropIfExists("ant_service");
+    if (Schema::hasTable('ant_service')) {
+      Schema::dropIfExists("ant_service");
+    }
   }
 }
