+++ database/migrations/2021_06_24_135933_create_ant_role_table.php
@@ -13,11 +13,13 @@ class CreateAntRoleTable extends Migration
    */
   public function up()
   {
-    Schema::create("ant_role", function (Blueprint $table) {
-      $table->string("role_id", 100)->primary();
-      $table->string("role_nama", 100);
-      $table->text("role_deskripsi")->nullable();
-    });
+    if (!Schema::hasTable('ant_role')) {
+      Schema::create("ant_role", function (Blueprint $table) {
+        $table->string("role_id", 100)->primary();
+        $table->string("role_nama", 100);
+        $table->text("role_deskripsi")->nullable();
+      });
+    }
   }
 
   /**
@@ -27,6 +29,8 @@ class CreateAntRoleTable extends Migration
    */
   public function down()
   {
-    Schema::dropIfExists("ant_role");
+    if (Schema::hasTable('ant_role')) {
+      Schema::dropIfExists("ant_role");
+    }
   }
 }
