+++ database/migrations/2021_10_07_091106_create_ant_dokter_table.php
@@ -13,17 +13,19 @@ class CreateAntDokterTable extends Migration
    */
   public function up()
   {
-    Schema::create("ant_dokter", function (Blueprint $table) {
-      $table->bigIncrements("id");
-      $table->string("dokter_nik", 100)->nullable();
-      $table->string("dokter_nama", 225);
-      $table->char("dokter_gender", 1)->nullable();
-      $table->date("dokter_birth_date")->nullable();
-      $table->string("dokter_address", 225)->nullable();
-      $table->string("dokter_jam_praktik", 225)->nullable();
-      $table->string("dokter_poli_id", 100)->nullable();
-      $table->timestamps();
-    });
+    if (!Schema::hasTable('ant_dokter')) {
+      Schema::create("ant_dokter", function (Blueprint $table) {
+        $table->bigIncrements("id");
+        $table->string("dokter_nik", 100)->nullable();
+        $table->string("dokter_nama", 225);
+        $table->char("dokter_gender", 1)->nullable();
+        $table->date("dokter_birth_date")->nullable();
+        $table->string("dokter_address", 225)->nullable();
+        $table->string("dokter_jam_praktik", 225)->nullable();
+        $table->string("dokter_poli_id", 100)->nullable();
+        $table->timestamps();
+      });
+    }
   }
 
   /**
@@ -33,6 +35,8 @@ class CreateAntDokterTable extends Migration
    */
   public function down()
   {
-    Schema::dropIfExists("ant_dokter");
+    if (Schema::hasTable('ant_dokter')) {
+      Schema::dropIfExists("ant_dokter");
+    }
   }
 }
